refactor: split anki-connect and stats-server resolvers into modules
- Break src/config/resolve/anki-connect.ts into focused resolvers under anki-connect/ (initialize, known-words, legacy, modern, shared)
- Break src/core/services/stats-server.ts into route groups under stats-server/ (analytics, integration, library, mining, static, shared support)
- Add route-group and resolver isolation tests
- Update docs/architecture/domains.md to reflect new module layout
