+++ src/core/services/stats-server/routes.ts
@@ -1,0 +1,5 @@
+export { registerStatsAnalyticsRoutes } from './analytics-routes.js';
+export { registerStatsIntegrationRoutes } from './integration-routes.js';
+export { registerStatsLibraryRoutes } from './library-routes.js';
+export { registerStatsMiningRoutes } from './mining-routes.js';
+export { registerStatsStaticRoutes } from './static-routes.js';
